% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.absorbance}
\alias{df.absorbance}
\title{df.absorbance}
\format{
A data frame with 1428 rows and 10 variables:
\describe{
\item{\code{Temperature}}{double Temperature in degCelcius}
\item{\code{Absorbance}}{double The absorbance measured using a spectrometer}
\item{\code{Experiment}}{character The experiment name}
\item{\code{RNA}}{character The RNA sequence}
\item{\code{Buffer}}{logical The buffer}
\item{\code{File}}{character The file path}
\item{\code{Cell}}{character The cell in a run}
\item{\code{Blank}}{character The blank the spectrometer uses}
\item{\code{Pathlength}}{double The pathlength of the cuvette in cm}
\item{\code{Wavelength}}{double The wavelength the absorbance was measured at in nm}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
df.absorbance
}
\description{
Absorbance data Znosko
}
\keyword{datasets}
